final fixes for task 3, requirements addition, proper lab structure
